[geometry] Made the constructor of Point<T> from Point<U> explicit.
Sometimes we want a non-trivial transformation when converting
from one representation of a point to another. An example
is PointDToPointU which uses the fact that its argument is
usually a point from the Mercator plane. The wrong usage
is easier to catch when implicit conversion between different point
types is prohibited.

An example of wrong usage that is prevented by this commit:
  // A function defined in another file.
  m2::PointU GetPoint()
  {
    m2::PointD pt;
    [...]
    return PointDToPointU(p2, kDefaultBits);
  }

  // Here we have forgotten that the result of GetPoint() is already converted.
  m2::PointU ptu = PointDToPointU(GetPoint(), kDefaultBits);