[coding] Default cache params in FileReader's constructor.
Now that we know all the places that depended on the
implicit params, fall back to the simple constructor.

The main problem here was that there used to be a component
in generator/intermediate_data.hpp where the same field could be
used both as a reader and as a writer via template metaprogramming.
This effectively forced the constructors of FileReader and
FileWriter to have the same parameter list.

Three default parameters in FileReader's constructor that could
be implicitly cast to one another only complicated the matters,
so once we have removed the |withException| param an extra
check was needed that all the constructors are now used as
intended.