[base] Osm::Id fixes.
The masks in the higher bits of our osm id encoding
used to be disjoint. After a refactoring, the encoding
was switched to packing 4 values into 2 bits and thus
the masks started to intersect. The checks were left
untouched, though, resulting in Relation passing tests
for being a Node and a Way.
This commit fixes the bug and introduces an enum to
store the osm id type, hoping to make things more clear.