[storage] Removing the deleted country from DiffManager.
Suppose that a user has one mwm and the app has been updated.
It now offers to download the diff, however instead
the user deletes the map and tries to download it again.
The download fails because according to the DiffManager it's
not the mwm but the mwmdiff that should be downloaded which
results in a size mismatch. Moreover, if the download is retried,
the diff is downloaded (because the MapOptions are correct now)
but it cannot be applied because the old mwm has already been removed.

The problem is that we do not account for the change in mwm status
since the start of the application (and, in particular, the DiffManager)
when taking the decision whether to download a diff or a full mwm.