[storage] Fixed a crash.
Consider the following scenario.
1. Diffs for A and B are downloaded but not applied.
   Currently it is possible because we do not delete a
   diff whose application process was cancelled.
2. The update button is clicked for both A and B, adding
   them both to the queue.
3. The cancel button is clicked for both A and B.

The cancellation of A sends a signal to the diff
manager and kicks the downloader to start applying B.
The diff manager receives the signal to cancel A and
resets the m_latestDiffRequest. Then, the cancel request
to B clears the queue but never reaches the diff manager
(because of the m_latestDiffRequest mismatch), and, once
the diff to B has been successfully applied, Storage cannot
remove B from queue because it has already been removed.