set_up_android.py for Windows configuration
The current implementation of set_up_android.py does not support Windows. In Windows each file is expected to have an file extension (opposed to Linux systems). A new function called `add_if_windows` adds only an extension if the operation system is `nt` (meaning that Windows specific modules are available).

If this change is not contributed, Windows users will face an error explained in issue #5258. The SDK and NDK will not be found. Manual changes to the file will then be necessary in order to make the setup work.