Leak in InputStreams
Same concept as before. However, this time the inputstreams were not closed at all. With the try-with-resources the inputstreams will also be closed and will prevent a leak from happening.

To come back to my last commit, the usage of 'memory leak' was wrong. This is a leak, not a memory leak. The garbage collector may eventually close the resources. However, some operating systems will not allow many simultaneous open resources.