[data] [storage] [generator] Removed the Vatican City mwm.
This commit removes "data/borders/Vatican City.poly" which
means that we won't generate the "Vatican City.mwm" anymore.
This mwm wasn't present in countries.txt and wasn't downloadable
via the app. The city itself is included in Italy_Lazio.mwm, so
the deleted mwm file was redundant anyway.

Several tests against this situation are added.