[editor] Postponed removing edits for a deleted map.
Editor used to subscribe to the now-deleted OnMapUpdated event
in the MwmSet. The semantics of OnMapUpdated(new,old) has always been
OnMapDeregistered(old)+OnMapRegistered(new) (the order is unspecified)
but the Editor treated it differently: i.e. in the Editor, an Update
was treated as a separate event non-equivalent to the combination of
Register and Deregister.

After the change in the implementation of the MwmSet, the OnMapUpdated
call is gone and, as per MwmSet's documentation, is replaced with
OnMapDeregistered+OnMapRegistered. This changes the behaviour of
the Editor.

Before, the Editor did nothing on a Registered event (all edits for the
existing maps were loaded at startup), reloaded all edits on an Update
event (including the edits for the non-updated maps), and deleted the
edits for the deleted map on a Deregistered event. The reasons for
deletion of the edits are unclear but presumably we wanted to warn
the user to not delete the map until the edits have been uploaded, and
also some bugs were observed when redownloading a map for which edits had
not been deleted.

Now, the Editor loads the edits for a newly Registered map and does
nothing when a map is Deregistered. The changes for a deleted map will be
deleted on the next startup or when another map is downloaded (or updated).
The message we show to the user is not changed and it still says that the
edits are going to be deleted.

Note that with current design we need the mwm file to be present if we
want to upload the changes. Therefore, the collection-deletion of edits
for showing them on map is not decoupled from the collection-deletion
for the purpose of uploading them to OSM.