[coding] Replace Uri with Url.
There is a lot of confusion regarding these terms ([0], [1], [2]).
The safer way seems to be using Uri everywhere in our code
but for now change it the other way around for consistency because
there is a lot less to rename this way.

This commit makes sure that the (in theory, the more general) class Uri
is not a part of the (in theory, more specific) namespace url.

[0] https://tools.ietf.org/html/rfc3305
[1] https://danielmiessler.com/study/difference-between-uri-url/
[2] https://stackoverflow.com/questions/176264/what-is-the-difference-between-a-uri-a-url-and-a-urn