[coding] [map] Refactored url's ForEachParam.
The bool value in the old version served two purposes
at once: it was used both to control the execution flow
and to indicate the parsing result. As such, the return
value of true meant "the needed key was found and the
callback returned true when executed with the corresponding value"
and the return value of false meant "either the needed key
was not found or the callback returned false even though the
key was present". This is too loaded semantically and is
a false optimization (since the parsed url params need to be stored
in a vector of pairs by the time of this call, early exit won't help much).