[search] Search in downloader by country names.
When a request to search in downloader arrives, we used to
only find features on the world map that match the request
and return the mwms that contain these features.

This commit mixes in the results of search directly in
the country tree (countries.txt), or, to be more precise, by
the translations of the names of the countries there (countries_names.txt).

This is not the most efficient implementation but hopefully
it isolated enough to make improvements easy and it was also
useful as an exploration where our current search APIs are lacking, for example

* The unnecessary std::string<->UniString conversions.
* Indexes such as MemSearchIndex pretending to be generic while in fact being
  tailored to a particular use-case.
* The difficulty of mixing search results from different sources.